Move common code to util.
